Avoid strange messages if PULSE_* props do not exist

+++ build-appimage.sh
@@ -52,10 +52,12 @@ cat > $TMPDIR/AppRun <<\EOF
 HERE=$(dirname $(readlink -f "${0}"))
 export LD_LIBRARY_PATH="${HERE}"/usr/lib:$LD_LIBRARY_PATH
 export XDG_DATA_DIRS="${HERE}/usr/share:${XDG_DATA_DIRS:-/usr/local/share:/usr/share}"
-export PULSE_COOKIE=`mktemp`
-trap 'rm -f $PULSE_COOKIE' EXIT
-xprop -root PULSE_COOKIE | cut -d '"' -f 2 | while read -N2 code; do printf "\x$code"; done > $PULSE_COOKIE
-export PULSE_SERVER=`xprop -root PULSE_SERVER | cut -d '"' -f 2 | cut -d '}' -f 2`
+if ! xprop -root PULSE_COOKIE | grep -q "no such atom"; then
+    export PULSE_COOKIE=`mktemp`
+    trap "rm -f $PULSE_COOKIE" EXIT
+    xprop -root PULSE_COOKIE | cut -d '"' -f 2 | while read -N2 code; do printf "\x$code"; done > $PULSE_COOKIE
+    export PULSE_SERVER=`xprop -root PULSE_SERVER | cut -d '"' -f 2 | cut -d '}' -f 2`
+fi
 export LIBVA_DRIVERS_PATH="${HERE}"/usr/lib
 "${HERE}"/usr/bin/spicy $@
 EOF
